Data Loading
Here is the data loading for the second assignment. When you download the data from http://www.vision.caltech.edu/Image_Datasets/Caltech256/ just add it to the same folder as where the notebook is. Its configured to run that way. Will need to divide the data into training, testing and validation sets next and then develop models.